Moved Input class into inputs module
